requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data collection
yfinance>=0.2.0
polygon-api-client>=1.0.0
newsapi-python>=0.2.6

# Trading and broker integration
alpaca-trade-api>=3.0.0
ccxt>=4.0.0

# Machine learning and analysis
scikit-learn>=1.1.0
scipy>=1.9.0
ta-lib>=0.4.0

# Natural language processing
nltk>=3.8
textblob>=0.17.0
vaderSentiment>=3.3.0

# Web framework (for dashboard)
flask>=2.3.0
flask-cors>=4.0.0

# Scheduling
apscheduler>=3.10.0

# Monitoring and logging
prometheus-client>=0.16.0
structlog>=23.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Production
gunicorn>=20.1.0
redis>=4.5.0
celery>=5.3.0

# Data visualization
plotly>=5.15.0
dash>=2.10.0

# Financial analysis
finrl>=0.3.0
pyfolio>=0.9.2
empyrical>=0.5.5